+++ bean-utils-library/src/test/java/com/hotels/beans/BeanUtilsTest.java
@@ -18,7 +18,6 @@ package com.hotels.beans;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 import static com.shazam.shazamcrest.matcher.Matchers.sameBeanAs;
@@ -36,13 +35,10 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import com.hotels.beans.error.InvalidBeanException;
-import com.hotels.beans.model.FieldTransformer;
 import com.hotels.beans.sample.FromFooSimple;
-import com.hotels.beans.sample.FromFooSimpleBooleanField;
 import com.hotels.beans.sample.immutable.ImmutableToFoo;
 import com.hotels.beans.sample.immutable.ImmutableToFooMissingCustomAnnotation;
 import com.hotels.beans.sample.immutable.ImmutableToFooSimple;
-import com.hotels.beans.sample.immutable.ImmutableToFooSimpleBoolean;
 import com.hotels.beans.transformer.Transformer;
 
 /**
@@ -114,26 +110,6 @@ public class BeanUtilsTest {
     }
 
     /**
-     * Test that the transformer function is applied earlier than the default value
-     */
-    @Test
-    public void testTransformerFunctionHasHigherPriorityThanDefaultValue() {
-        //GIVEN
-        BeanUtils beanUtils = new BeanUtils();
-        FromFooSimpleBooleanField fromFooSimpleNullFields = createFromFooSimpleNullFields();
-        FieldTransformer<Boolean, Boolean> nullToTrue =
-                new FieldTransformer<>("work", aBoolean -> aBoolean == null || aBoolean);
-
-        //WHEN
-        ImmutableToFooSimpleBoolean transformed = beanUtils.getTransformer()
-                .withFieldTransformer(nullToTrue)
-                .transform(fromFooSimpleNullFields, ImmutableToFooSimpleBoolean.class);
-
-        //THEN
-        assertTrue(transformed.getWork());
-    }
-
-    /**
      * Creates the parameters to be used for testing the default transformation operations.
      * @return parameters to be used for testing the default transformation operations.
      */
@@ -182,11 +158,4 @@ public class BeanUtilsTest {
         return new FromFooSimple(NAME, ID);
     }
 
-    /**
-     * Creates a {@link FromFooSimpleBooleanField} instance with null field.
-     * @return the {@link FromFooSimpleBooleanField} instance.
-     */
-    private FromFooSimpleBooleanField createFromFooSimpleNullFields() {
-        return new FromFooSimpleBooleanField();
-    }
 }
